Sort datapoints on insert
Summary:
Sorting was taking a significant amount of time on scrapes, even though most of the time things are already in order. Instead, insert datapoints in sorted order (which usually requires no op since datapoints usually come in order) and then we don't have to do any work on scrape. Additionally scrapes are almost always longer than receives so it's better to push the work there if we can.

Reasoning:
1. Scrape times are usually longer so we want to minimize that time as much as possible
2. Sorting on scrape used a significant amount of CPU time and memory (verified via profiling)
3. Datapoints in all but the strangest deployments will generally arrive in order
4. Sorting on insert if they arrive out of order will generally require a lot less work with this assumption

Reviewed By: rckclmbr

Differential Revision: D22989764

fbshipit-source-id: 9908003a01703e27824d65cbad3bba3f49a1d32b